add error handling for server connection
