+++ assignment/client/src/feature/page/Home.js
@@ -11,9 +11,9 @@ export default function Home(props) {
   const updateQuestionsList = async () => {
     try {
       let data = await getQuestions();
-      console.log(data);
       setQuestions(data);
     } catch (error) {
+      setQuestions([]);
       console.log(error);
     }
   };
