+++ assignment/client/src/feature/questions/QuestionTable.js
@@ -1,6 +1,14 @@
 import React, { Fragment } from "react";
 import QuestionTableRow from "../questions/QuestionTableRow";
-import { Table, Thead, Tbody, Tr, Th, TableContainer } from "@chakra-ui/react";
+import {
+  Table,
+  Thead,
+  Tbody,
+  Tr,
+  Th,
+  TableContainer,
+  Text,
+} from "@chakra-ui/react";
 
 export default function QuestionTable(props) {
   return (
@@ -16,18 +24,22 @@ export default function QuestionTable(props) {
           </Tr>
         </Thead>
         <Tbody>
-          {props.questions.map((data) => (
-            <Fragment key={data.id}>
-              <QuestionTableRow
-                {...data}
-                attemptQuestion={props.attemptQuestion}
-                updateQuestionsList={props.updateQuestionsList}
-                questions={props.questions}
-              />
-            </Fragment>
-          ))}
+          {props.questions.length !== 0 &&
+            props.questions.map((data) => (
+              <Fragment key={data.id}>
+                <QuestionTableRow
+                  {...data}
+                  attemptQuestion={props.attemptQuestion}
+                  updateQuestionsList={props.updateQuestionsList}
+                  questions={props.questions}
+                />
+              </Fragment>
+            ))}
         </Tbody>
       </Table>
+      {props.questions.length === 0 && (
+        <Text color="red">Server encountered error fetching questions</Text>
+      )}
     </TableContainer>
   );
 }
